modularise ball, pikachu and shape functions
